I'm replacing the existing RS274X class (that uses polygons) to this one that uses OpenCascade TopoDS_Face objects for intersections.  It works for all examples we have so far except test1.gbr.  That example has a tangent intersection (touch at single point) that occurs at two separate points between two polygons.  I can't figure out how to allow for that but I believe that this version is more functional that the existing version.
I had started to go down the path of rendering the PCB into a raster image a view to performing an edge detection algorithm on the raster image. I got to the point where the raser image is generated (except for filling in the filled areas).  I stopped at this point because I had repaired enough of the faults in the raw traces data that the OpenCascade intersection code began working for all the various cases.

The OpenCascade routines do not detect an area that is completely inside an area so there are examples where short traces are left as polygons inside larger trace areas.  Although this may require some manual cleanup, it is at least working.

I also added code to reorder the sketch objects into counter-clockwise sketches where possible.  The intent being to allow a pocketing operation to be applied around all the trace sketches to remove copper between traces (if needed).  If this is not needed then the consistent orientation of the traces doesn't do any harm.

This routine will still be inefficient (slow) but I've played with it long enough to not care about this.  I just want it to work at this point.