Removed the 'imbue("C")' line for std::ostringstream so that the locale settings will be used for floating point conversion. This should fix Issue 197 in HeeksCNC.
Fixed problem where the number of decimal points was one
less than required when the value was negative.  i.e. I forgot
to allow for the minus sign when calculating how many
digits were required to reproduce the current tolerance
value.  Without this, sketches that are written out will not
join up again when used by the OpenCascade library
routines (which use the tolerance value everywhere).