Added some checks for a NULL HeeksObj* pointer following a GetIDObject() call.  This was causing some problems as per Issue 217 in HeeksCNC.
Changed definition of Ttc() and Ctt() so that they convert what
they receive into wxString objects and return a utf8 string.
This allows for more languages while not compilcating an
ASCII data file for English settings.  NOTE: This change meant
passing original wxString objects into these conversions
so that the text strings were kept in their Unicode form for
as long as possible.  i.e. we want to keep text within
wxString objects so that language constraints can be handled
by the wxString class and convert them to multi-byte
character strings using the UTF-8 encoding scheme.

Made lots of changes to ensure that utf8_str() strings were
being used in the interactions to/from Properties as well
as to/from XML data.

These UTF-8 changes are being made for Issue 280 in HeeksCAD